feat(planner): enable row fetch by default
