+++ src/query/sql/src/planner/binder/mod.rs
@@ -72,6 +72,7 @@ pub use internal_column_factory::INTERNAL_COLUMN_FACTORY;
 pub use location::get_storage_params_from_options;
 pub use location::parse_storage_params_from_uri;
 pub use location::parse_uri_location;
+pub use merge_into::target_table_position;
 pub use merge_into::MergeIntoType;
 pub use scalar::ScalarBinder;
 pub use scalar_common::*;
